etl-data-quality/requirements.txt
apache-airflow-providers-postgres[amazon]
psycopg2-binary
pandas